requirements.txt
# Flask API Requirements
flask>=3.1.1
google-adk>=1.3.0
google-genai>=1.20.0
python-dotenv>=1.1.0
asyncio>=3.4.3

# Streamlit App Requirements
streamlit>=1.46.0
requests>=2.32.4
pandas>=2.3.0
plotly>=6.1.2

# Development and Testing
pytest>=8.4.1
black>=25.1.0
flake8>=7.2.0 